Small cleanups: consistent types, deduplicate flag, fix capitalization
- opcodes.h: is_absjump parameter uint32_t → uint8_t to match is_jump
- codegen_statement.c: remove redundant early CO_VARKEYWORDS set
- scanner.c: lowercase "unexpected input char" to match all other messages

+++ opcodes.h
@@ -158,7 +158,7 @@ enum format_value_arg {
   FORMAT_VALUE_FMT_SPEC = 0x04,
 };
 
-static inline bool is_absjump(uint32_t opcode)
+static inline bool is_absjump(uint8_t opcode)
 {
   switch (opcode) {
   case OPCODE_JUMP_ABSOLUTE:
